* Refactor sim bindings (#10)

* update gitignore to remove editor config

* move and rename some files

* update godot version for real?

* refactor defines into enum

* rename fcsim piece enum

* make enum order consistent between render and sim

* make switch order consistent

* use diameter instead of radius when rendering circles

* fix diameters i forgot to convert from radii

* add and use addPiece function

* fix broken circle joints

* update gitignore to remove .vscode

* update gitignore to remove .vscode

* pytest: system tests and simulation matching FC (66% pass) (#11)

* import dataset

* test rig skeleton

* delete binary

* gitignore run_single_design

* run_single_design

* helpful comments

* basic optimizations

* tests but no run, no real data

* run subprocess

* get_design

* parse design struct

* local design cache

* change blocks to vector

* use uint16_t

* run executable, test for retcode 0

* Revert "change blocks to vector"

This reverts commit ac8a87f4547463b4765fbd710fdf71aee5b9abce.

* test for correct tick count

* note about 8-bit ints

* refactor run_design to new file

* remove placeholder tests

* infrastructure for valgrind

* propagate user comment

* unify error messages

* more helpful shell error message

* test valgrind

* fix valgrind check

* pass rate report

* more colors

* enable hardfloat, s794

* increase default ticks budget

* fix radius on SC/DC

* better descriptions

---------

+++ src/sim/fcsim.h
@@ -28,7 +28,7 @@ inline bool type_is_player(int t) {
 }
 
 struct fcsim_block_def {
-	uint8_t type;
+	uint16_t type;
 	uint16_t id;
 	double x, y;
 	double w, h;
@@ -121,6 +121,7 @@ template <typename T> void delete_all(std::vector<T*>& vec) {
 }
 
 bool fcsim_in_area(const fcsim_block_def& bdef, const fcsim_rect& area);
+
 bool fcsim_is_solved(const std::shared_ptr<ft_sim_state> sim, const ft_design_spec& spec);
 
 #endif
